feat: draw Unicode text with font stacks
Historically we've drawn text using sprite fonts representing text in
the user ACP or later in UTF-8. Any unrecognized codepoints were
represented as composite tofu glyphs in a smaller font.

This introduces a font stack system with dynamic atlas textures,
rendering text with font fallback and more complete Unicode shaping and
traversal.

It leverages a number of libraries:

 * FreeType for vector font rasterisation,
 * libfribidi for text segmentation by text script,
 * HarfBuzz for text shaping,
 * stb_rect_pack for packing glyph atlas textures,
 * ztd.text for transcoding text.

This change drops the previous tofu support, assuming that enough fonts
are present to support the commonly used scripts.

+++ engine/common/console.cpp
@@ -782,7 +782,7 @@ void conInputHandler_c::ConInputKeyEvent(int key, int type)
 		case KEY_TAB:
 			if (_con->input.len) {
 				std::string comp = _con->input.buf;
-				int	compLen = comp.size();
+				int	compLen = (int)comp.size();
 
 				// Build match list
 				int num = 0;
